add base profile viewer class
