+++ jdaviz/configs/__init__.py
@@ -1,6 +1,7 @@
 from .cubeviz import *  # noqa
+from .default import *  # noqa
+from .imviz import *  # noqa
+from .mosviz import *  # noqa
+from .rampviz import * # noqa
 from .specviz import *  # noqa
 from .specviz2d import *  # noqa
-from .default import *  # noqa
-from .mosviz import *  # noqa
-from .imviz import *  # noqa
