+++ jdaviz/configs/rampviz/__init__.py
@@ -1,0 +1,2 @@
+from .plugins import *  # noqa
+from .helper import Rampviz  # noqa
