+++ jdaviz/utils.py
@@ -19,6 +19,7 @@ from glue.config import settings
 from glue.core import BaseData
 from glue.core.exceptions import IncompatibleAttribute
 from glue.core.subset import SubsetState, RangeSubsetState, RoiSubsetState
+from glue_astronomy.spectral_coordinates import SpectralCoordinates
 from ipyvue import watch
 
 from jdaviz.core.validunits import check_if_unit_is_per_solid_angle
@@ -517,7 +518,7 @@ def get_subset_type(subset):
     Returns
     -------
     subset_type : str or None
-        'spatial', 'spectral', or None
+        'spatial', 'spectral', 'temporal', or None
     """
     if not hasattr(subset, 'subset_state'):
         return None
@@ -530,7 +531,40 @@ def get_subset_type(subset):
     if isinstance(subset.subset_state, RoiSubsetState):
         return 'spatial'
     elif isinstance(subset.subset_state, RangeSubsetState):
-        return 'spectral'
+        # look within a SubsetGroup, or a single Subset
+        subset_list = getattr(subset, 'subsets', [subset])
+
+        for ss in subset_list:
+            if hasattr(ss, 'data'):
+                ss_data = ss.data
+            elif hasattr(ss.att, 'parent'):
+                # if `ss` is a subset state, it won't have a `data` attr,
+                # check the world coordinate's parent data:
+                ss_data = ss.att.parent
+            else:
+                # if we reach this `else`, continue searching
+                # through other subsets in the group to identify the
+                # subset type:
+                continue
+
+            # check for a spectral coordinate in FITS WCS:
+            wcs_coords = (
+                ss_data.coords.wcs.ctype if hasattr(ss_data.coords, 'wcs')
+                else []
+            )
+
+            has_spectral_coords = (
+                any(str(coord).startswith('WAVE') for coord in wcs_coords) or
+
+                # also check for a spectral coordinate from the glue_astronomy translator:
+                isinstance(ss_data.coords, SpectralCoordinates)
+            )
+
+            if has_spectral_coords:
+                return 'spectral'
+
+        # otherwise, assume temporal:
+        return 'temporal'
     else:
         return None
